Add post-scan action groups, new actions, and web UI refactoring
Backend:
- New action groups architecture: PostScanActionGroup + PostScanActionGroupScanner (join table)
- PostScanManager runs matching groups in parallel, actions within group sequentially
- 6 new post-scan actions: Telegram, Email, DataEnrichment, Validation, Aggregation, DatabaseQuery
- WindowPaste action: paste into specific window by title with AttachThreadInput activation
- ClipboardPasteAction: retry logic for clipboard opening, async delays, CancellationToken support
- ValidationAction: new 'format' mode replacing built-in parser validation
- SimpleBarcodeParser: removed hardcoded IsValid, validation moved to ValidationAction
- ScanResult: added Metadata dictionary for enrichment results
- ScanTemplateHelper: shared template formatting utility
- EF migration: ActionGroups with clean schema (no FK constraints for SQLite compatibility)

Web UI:
- Refactored monolithic index.html into modular CSS/JS structure (12 files)
- Group cards replacing flat action table
- Group editor modal with scanner binding checkboxes
- Action editor modal within groups
- Action type descriptions in editor
- New action type options and settings for all new actions

Tests:
- 173/173 tests passing
- New test files for all 6 new actions
- Parser tests updated for validation architecture change

+++ src/Data/Entities/LogRecord.cs
@@ -1,10 +1,34 @@
 namespace ScanBridge.Data.Entities;
 
+/// <summary>
+/// Сущность записи лога в базе данных SQLite.
+/// Хранит информацию о событиях, происходящих в системе.
+/// </summary>
 public class LogRecord
 {
-    public long Id { get; set; }
+    /// <summary>
+    /// Уникальный идентификатор записи.
+    /// </summary>
+    public int Id { get; set; }
+
+    /// <summary>
+    /// UTC-метка времени события.
+    /// </summary>
     public DateTime Timestamp { get; set; }
+
+    /// <summary>
+    /// Уровень логирования: INF, WRN, ERR, DBG, FTL, VRB.
+    /// Ограничение: до 10 символов.
+    /// </summary>
     public string Level { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Текст сообщения лога. Ограничение: до 4000 символов.
+    /// </summary>
     public string Message { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Текст исключения (если было). Может быть null.
+    /// </summary>
     public string? Exception { get; set; }
 }
